requirements.txt
Flask==3.0.3
langchain==0.2.14
langchain-core==0.2.39
langchain-community==0.2.12
langchain_groq==0.1.9
pandas==2.0.2
python-dotenv==1.0.1
